requirements.txt
python-dotenv
pydantic
tqdm
loguru

requests
beautifulsoup4
lxml
playwright

pandas
numpy

sentence-transformers
faiss-cpu
chromadb

langchain
langchain-community
langchain-text-splitters
llama-index

google-generativeai

fastapi
uvicorn

scikit-learn

orjson

gunicorn